sprint1/5-separate-swagger-file-and-contents
